week02/requirements.txt
python-dotenv
requests
lxml
matplotlib
pandas
numpy
seaborn
plotly
drawsvg